renames classes
